Move pillars.ts to frontend until we have a folder for things like this
